requirements.txt
black==19.3b0
coveralls==1.7.0
flask==1.0.2
flask-sqlalchemy==2.3.2
flask-script==2.0.6
flask-migrate==2.4.0
flake8==3.7.7
gunicorn==19.9.0
psycopg2-binary==2.7.7
pytest==4.4.1
pyjwt==1.7.1
tox==3.9.0
webargs==5.3.0
